requirements.txt
streamlit==1.16.0
PyPDF2==3.0.0
google-generativeai==0.1.0
pinecone-client==2.2.1
python-dotenv